Climber code re-write for testing on Friday
This is a re-write of the climbing code. It is the simplest possible code (just using motor power).
